requirements.txt
PyLD>=2.0.4
cryptography>=41.0.0
PyNaCl>=1.5.0
base58>=2.1.1
qrcode[pil]>=7.4.0
Pillow>=10.0.0
